+++ src/plugins/rx3270/pluginmain.cc
@@ -135,8 +135,8 @@
 	string			* get_host_charset(void);
 	string			* get_display_charset(void);
 
-	const char  	* asc2ebc(unsigned char *str, size_t sz = -1);
-	const char	 	* ebc2asc(unsigned char *str, size_t sz = -1);
+	const char  	* asc2ebc(unsigned char *str, int sz = -1);
+	const char	 	* ebc2asc(unsigned char *str, int sz = -1);
 
     int				  quit(void);
 
@@ -761,12 +761,12 @@ int	plugin::print(void)
 	return lib3270_print(hSession);
 }
 
-const char * plugin::asc2ebc(unsigned char *str, size_t sz)
+const char * plugin::asc2ebc(unsigned char *str, int sz)
 {
 	return lib3270_asc2ebc(hSession,str,sz);
 }
 
-const char * plugin::ebc2asc(unsigned char *str, size_t sz)
+const char * plugin::ebc2asc(unsigned char *str, int sz)
 {
 	return lib3270_ebc2asc(hSession,str,sz);
 }
